backend/pyproject.toml
[project]
name = "fiscateste-backend"
version = "0.1.0"
description = "FastAPI backend with JWT auth and PostgreSQL"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt<5",
    "email-validator>=2.2.0",
    "fastapi>=0.116.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
